sensors/classification/requirements.txt
tensorflow==2.2.0
keras==2.4.0
scikit-learn==0.24.0
numpy==1.18.5
matplotlib
colour
pandas
mpld3
pydot
six
pyyaml==5.4.1
seaborn==0.11.2
protobuf==3.20.*